Finance/requirements.txt
Flask==2.2.1
python_dateutil==2.8.2
requests==2.27.1
